scrollPercentY should return a value between 0 and 1
In chrome when you scroll outside of the viewport either above or below the page:

- chrome will **not** continue to fire scroll events
- chrome will report a scrollTop of either 0 or the maximum possible scrollable height.

i.e. there is a floor and a ceiling in play here.

As such our scrollPercentY function can only return a float between 0 and 1

In webkit this isn't the case

- webkit will **continue** to fire scroll events during the "bounce"
- webkit will return a negitive number for scrollTop if bouncing up, and will return a larger number than the maximum possible scrollable height if bouncing below the page.

As such our scrollPercentY function can return a float that's negative and also great than 100%.

But what i think is really causing the issue of displaying the promos is that the scrollPercentY function safari can return a non float value because of this.

e.g.

If you scroll above the page, it will go into negative pixels, because the page isn't really scrollable the calculation will be -5/0 where -5 is the current scroll position and 0 is the scrollable heigh (i.e. no scroll available when the menu is visible) this returns `-infinity` and when it's 10/0 it will return `infinity`.

Also, when if it returns 0 scroll, it's 0/0 which returns NaN.